pyproject.toml
[build-system]
requires = [
    "setuptools>=64.0.0",
    "wheel",
    "setuptools_scm[toml]>=7.0",
    "check-manifest"
]

[project]
name = "xarray-eop"
authors = [{"name" = "Vincent Levasseur"}]
dynamic = ["version"]
readme = "README.md"
description = "Xarray backends to read Sentinel products (SAFE and zarr formats)"
requires-python = ">=3.11"
dependencies = [
    "click",
    "colorlog",
    "dask",
    "fsspec",
    "matplotlib",
    "netcdf4>=1.6.5",
    "h5netcdf>=1.3.0",
    "scipy",
    "s3fs >= 2023.10.0",
    "universal_pathlib",
    "xarray >= 2023.10.1",
    "xarray-datatree >= 0.0.14",
    "zarr",
]
license = {file = "LICENSE"}

[project.optional-dependencies]
notebook = [
    "jupyter"
]
doc = [
    "pickleshare",
    "sphinx",
    "myst-nb",
    "sphinx-book-theme"
]
tests = [
    "pytest",
    "pytest-cov",
]
[tool.setuptools_scm]
write_to = "xarray_eop/_version.py"
write_to_template = '''
# Do not change! Do not track in version control!
__version__ = "{version}"
'''

[tool.flit.sdist]
exclude = ["conf/"]

[tool.setuptools.packages.find]
exclude = ["conf/"]

[tool.pytest.ini_options]
markers = [
    "integration",
    "unit",
]
[project.entry-points."xarray.backends"]
sentinel-3 = "xarray_eop.xarray_backends:Sentinel3Backend"
eop = "xarray_eop.xarray_backends:EOPBackend"
